Criação e testes do handler que cria um novo filme

+++ CatalogoFilmesSeries.Application/UseCases/Filmes/AdicionarResponse.cs
@@ -10,7 +10,7 @@ public sealed record AdicionarResponse(
     string UrlImagem,
     
     double AvaliacaoImdb,
-    int PopularidadeImdb,
+    double PopularidadeImdb,
     
     DateTime DataInclusao
 );
